re organize variable fetcher
