+++ frontend/taipy-gui/base/src/components/Taipy/components/BindingInput.tsx
@@ -5,9 +5,10 @@ import Box from "@mui/material/Box";
 import MenuItem from "@mui/material/MenuItem";
 import FormControl from "@mui/material/FormControl";
 import Select, { SelectChangeEvent } from "@mui/material/Select";
+import { getFunctionList, getVarList } from "../utils";
 
 const common = (props: TaipyElementInput, bindingInfo: string[][]) => {
-    const { onChange, value, defaultValue } = props;
+    const { onChange, value } = props;
     const handleChange = (event: SelectChangeEvent) => {
         if (event.target.value === "") {
             onChange(undefined);
@@ -33,17 +34,7 @@ const common = (props: TaipyElementInput, bindingInfo: string[][]) => {
 export const BindingInput = (props: TaipyElementInput) => {
     const app = useStore((state) => state.app);
     const data = useMemo(() => {
-        if (!app) {
-            return [];
-        }
-        const varData = app.getDataTree();
-        if (!varData) {
-            return [];
-        }
-        const varDataModule = varData[app.getContext()];
-        return Object.keys(varDataModule)
-            .filter((key: string) => !key.includes("chlkt") && !key.toLowerCase().includes("taipy"))
-            .map((key: string) => [varDataModule[key].encoded_name, key]);
+        return getVarList(app);
     }, [app]);
     return common(props, data);
 };
@@ -51,13 +42,7 @@ export const BindingInput = (props: TaipyElementInput) => {
 export const FunctionBindingInput = (props: TaipyElementInput) => {
     const app = useStore((state) => state.app);
     const data = useMemo(() => {
-        if (!app) {
-            return [];
-        }
-        return app
-            .getFunctionList()
-            .filter((key: string) => !key.includes("chlkt") && !key.toLowerCase().includes("taipy"))
-            .map((key: string) => [key, key]);
+        return getFunctionList(app);
     }, [app]);
     return common(props, data);
 };
